Increase LJ cutoff to 4*sigma
[#83064836]

The cutoff was reduced to 2 two years ago. Increasing it again solves poor *apparent* energy conservation caused by the fact that we don't cutoff the calculation of LJ potential, though we do cutoff the calculation on LJ force. (The LJ potential drops off fast; at 4*sigma it is negligible.)

(We could just cutoff the potential calculation. However in general the LJ cutoff was too aggressive as there is a clear step change in the force when atoms approach and reach the cutoff distance. Furthermore, with 2 more years of mobile device amd JS engine improvements it's worth risking a slight performance hit to use the more accurate cutoff.)