Settable chem reactions update interval
[#74691454]

When the chemical-reactions plugin is run every 10fs and bond-vibration periods are smaller than that, the plugin fails to observe the bond potential energy exceeding the bond dissociation energy. This hasn't mattered for more-qualitative models that have many molecules, but interactives that allow a user to examine bond formation or breakage more quantitatively may need for the plugin to run every integration step.