Account for LJ-potential difference between bond creation and destruction
[#74691454]

The engine does not compute the LJ force between bonded pairs. Thus, when a bond forms, a certain amount of potential energy "disappears". The previous solution just added this missing energy as kinetic energy, however, this meant that the total energy added when a bond was created exceeded the bond's dissociation energy. The new solution is to remember the LJ potential when the bond was created, include the remembered value in the system's overall potential energy, and absorb or release kinetic energy (according to the LJ potential difference at creation vs. at destruction) when the bond is broken or transferred. (This achieves the same effect as if the engine were calculating the LJ potential of the bond; in retrospect a better solution may be to require the engine to compute LJ potential of bonded pairs when the chemical reactions plugin is used.)