“official” hasPlayed flag is a model.property
[#84205630]

This fixes a bug whereby sensor-connector models wouldn’t turn on the analyze data button in CODAP. ExportController was relying on model.hasPlayed, not model.properties.hasPlayed. For implementation reasons, model.hasPlayed is defined on the MD2D model object, but it is not so defined on non-PlaybackSupport models like sensor-connector. Therefore one should use the official API (model.properties.hasPlayed) instead.