Fix incorrect enabling of analyze-data button
[#84301302]

Previously, playback controller was querying state related to canExportData before the export controller had a chance to act on that state. Fixed by dividing the exportController’s events into two (one indicating merely that we’re in CODAP, the other that the exportController has finished initialization for the current model and is now in a consistent state), and arranging for the playback control to (indirectly) consume the second event instead of the first.