Fix bug where removeModelData on step -1 throws error
When we update a property, we throw an invalidation event, which calls
removeModelDataAfterStepPointer in DataSet. DataSet then resets all its
data arrays to be the length of the current step, and then re-adds the
current point.

The problem is that if we call this function when the model is still
loading, we are on step "-1". DataSet tries to set its array lengths
to -1, which throws an error.

This happens any time we have an interactive with graphs, and we set
model properties in the onLoad function.

[#84394086]