Zero acceleration in pinAtoms, switch to just one call.
This replaces the zeroing of acceleration in pinAtoms as it is needed
for minimizeEnergy.

We can safely call pinAtoms just once, at the start of the integration
loop. This fixes the original issue essentially by just calling pinAtoms
earlier, because we now don't assume that accelerations can't have been
changed between integrations.

[#83957704]