Fix iframe errors after model / interactive reload
[#85083546]

It was caused by the fact that the old iframe phone was still connected
and it was trying to process all the incoming messages. However its
iframe wasn't in DOM anymore, so its .contentWindow was equal to null.

Make sure in Lab that old iframe phone is disconnected when needed.
I've added custom 'destroyed' jQuery event that is triggered when
element is removed from DOM. I've used it to disconnect phone when
its iframe is removed from DOM.